feat: dry_run release option

+++ lib/git_ops/version_replace.ex
@@ -3,19 +3,19 @@ defmodule GitOps.VersionReplace do
   Functions that handle the logic behind replacing the version in related files.
   """
 
-  @spec update_mix_project(module, String.t(), String.t()) :: :ok | {:error, :bad_replace}
-  def update_mix_project(mix_project, current_version, new_version) do
+  @spec update_mix_project(module, String.t(), String.t()) :: String.t() | {:error, :bad_replace}
+  def update_mix_project(mix_project, current_version, new_version, opts \\ []) do
     file = mix_project.module_info()[:compile][:source]
 
-    update_file(file, "@version \"#{current_version}\"", "@version \"#{new_version}\"")
+    update_file(file, "@version \"#{current_version}\"", "@version \"#{new_version}\"", opts)
   end
 
-  @spec update_readme(String.t(), String.t(), String.t()) :: :ok | {:error, :bad_replace}
-  def update_readme(readme, current_version, new_version) do
-    update_file(readme, ", \"~> #{current_version}\"", ", \"~> #{new_version}\"")
+  @spec update_readme(String.t(), String.t(), String.t()) :: String.t() | {:error, :bad_replace}
+  def update_readme(readme, current_version, new_version, opts \\ []) do
+    update_file(readme, ", \"~> #{current_version}\"", ", \"~> #{new_version}\"", opts)
   end
 
-  defp update_file(file, replace, pattern) do
+  defp update_file(file, replace, pattern, opts) do
     contents = File.read!(file)
 
     new_contents = String.replace(contents, replace, pattern)
@@ -23,9 +23,11 @@ defmodule GitOps.VersionReplace do
     if new_contents == contents do
       {:error, :bad_replace}
     else
-      File.write!(file, new_contents)
+      unless opts[:dry_run] do
+        File.write!(file, new_contents)
+      end
 
-      :ok
+      String.trim(new_contents, contents)
     end
   end
 end
